@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import store from './store';
 import App from './App';
@@ -7,11 +7,12 @@ import { MemoryRouter, BrowserRouter } from 'react-router-dom';
 
 const Router = process.env.NODE_ENV === 'test' ? MemoryRouter : BrowserRouter;
 
-ReactDOM.render(
-  <Provider store={store}>
-    <Router>
+const root = createRoot(document.getElementById('app'));
+
+root.render(
+  <Router>
+    <Provider store={store}>
       <App />
-    </Router>
-  </Provider>,
-  document.getElementById('app')
+    </Provider>
+  </Router>
 );
